+++ src/PixelWindowSystem/IPixelWindowAppManager.cs
@@ -17,7 +17,7 @@
         /// Update function, for updating non-render data every frame (e.g. handling input)
         /// </summary>
         /// <param name="frameTime">The time in ms since the last frame</param>
-        void Update(double frameTime);
+        void Update(float frameTime);
 
         /// <summary>
         /// Fixed update function, for changing non-render data at a fixed rate independent of rendering (e.g. 50hz)
@@ -26,7 +26,7 @@
         /// The fixed timestep configured on the <see cref="PixelWindow"/> in ms. This can and
         /// should be used as the deltatime value for any physics calculations for example.
         /// </param>
-        void FixedUpdate(double timeStep);
+        void FixedUpdate(float timeStep);
 
         /// <summary>
         /// Render function, for updating the raw pixel data
@@ -40,6 +40,6 @@
         /// Make sure to call <see cref="PixelData.Clear"/> at the start of the render function if you want to clear
         /// the screen before you start rendering!
         /// </remarks>
-        void Render(PixelData pixelData, double frameTime);
+        void Render(PixelData pixelData, float frameTime);
     }
 }
